rename - #15: ManagerDetails, ManagerDetailsService 분리
